+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2023-08-01
+
+### Changed
+
+- BREAKING: Don't require specifying update type for watchers with labels (#19)
+
+### Added
+
+- Allow watchers without an ecto schema (thanks @frerich / #18)
+
 ## [0.7.0] - 2023-07-31
 
 ### Changed